requirements.txt
Django
gunicorn
whitenoise
psycopg2-binary
python-dotenv
dj-database-url
djangorestframework